pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Akash Meshram <6388495akash@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
numpy = "^2.0.0"
python-dotenv = "^1.0.1"
python-socketio = "^5.11.3"
websocket-client = "^1.8.0"
requests = "^2.32.3"
cryptography = "^42.0.8"
asyncio = "^3.4.3"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:run"